% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anscombise.R
\name{anscombise}
\alias{anscombise}
\title{Create new versions of Anscombe's quartet}
\usage{
anscombise(x, which = 1)
}
\arguments{
\item{x}{A numeric matrix or data frame.  Each column contains observations
on a different variable.  Missing observations are not allowed.}

\item{which}{An integer in {1, 2, 3, 4}.  Which of Anscombe's dataset to
use.  Obviously, this makes very little difference.}
}
\value{
A dataset with the same format as \code{x}.  The returned dataset has the
same summary statistics as those in \code{stats}, except perhaps for the sample
size.
}
\description{
Modifies a dataset \code{x} so that it shares sample summary statistics with
\code{\link[datasets:anscombe]{Anscombe's quartet}}.
}
\examples{
new_faithful <- anscombise(datasets::faithful)
plot(new_faithful)
}
