% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anscombiser-package.R
\docType{package}
\name{anscombiser-package}
\alias{anscombiser-package}
\alias{anscombiser}
\title{anscombiser: Create Datasets with Identical Summary Statistics}
\description{
Anscombe's quartet (Anscombe, 1973) are a set of four two-variable datasets
that have several common summary statistics but which have very different
joint distributions. This becomes apparent when the data are plotted, which
illustrates the importance of using graphical displays in Statistics.  This
package enables the creation of datasets that have identical marginal sample
means and sample variances, sample correlation, least squares regression
coefficients and coefficient of determination.  The user supplies an initial
dataset, which is shifted, scaled and rotated in order to achieve target
summary statistics.  The general shape of the initial dataset is retained.
The target statistics can be supplied directly or calculated based on a
user-supplied dataset.
}
\details{
The main functions in \code{anscombiser} are
\itemize{
\item \code{\link{anscombise}}, which modifies a user-supplied dataset so that it shares
sample summary statistics with Anscombe's quartet.
\item \code{\link{mimic}}, which modified a user-supplied dataset so that is shares
sample summary statistics with another user-supplied dataset.
}

See \code{vignette("intro-to-anscombiser", package = "anscombiser")} for
an overview of the package.

The functions \code{\link{anscombise_gif}} and \code{\link{mimic_gif}} enable the animation of
several datasets produced by \code{\link{anscombise}} and \code{\link{mimic}}, respectively.
}
\references{
Anscombe, F. J. (1973). Graphs in Statistical Analysis.
The American Statistician 27 (1): 17–21.
\doi{10.1080/00031305.1973.10478966}
}
\seealso{
\code{\link{anscombise}}, \code{\link{mimic}}, \code{\link{anscombise_gif}} and \code{\link{mimic_gif}}.
}
\author{
\strong{Maintainer}: Paul J. Northrop \email{p.northrop@ucl.ac.uk} [copyright holder]

}
