% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anscombiser.R
\docType{package}
\name{anscombiser}
\alias{anscombiser}
\title{anscombiser: Create datasets with identical summary statistics}
\description{
Describe
}
\details{
Add
}
\references{
Steph Locke and Lucy D'Agostino McGowan (2018). datasauRus:
Datasets from the Datasaurus Dozen. R package version 0.1.4.
\url{https://CRAN.R-project.org/package=datasauRus}
}
\seealso{
\link{anscombise} and \link{mimic}
}
