% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdata.R
\name{mapdata}
\alias{mapdata}
\title{Extract longitude and latitude values}
\usage{
mapdata(region = ".", map = "world", exact = FALSE, ...)
}
\arguments{
\item{region}{Passed to \code{\link[maps:map]{map}} as the argument \code{regions}.}

\item{map}{Passed to \code{\link[maps:map]{map}} as the argument \code{database}}

\item{exact}{The argument \code{exact} passed to the \code{\link[maps:map]{map}} function.}

\item{...}{Additionl arguments to be passed to \code{\link[maps:map]{map}}.}
}
\value{
A dataframe with two columns: \code{long} and \code{lat} for longitude and
latitude.
}
\description{
Extracts longitude and latitude values for a particular region from the
world map supplied by the maps package.
}
\section{Examples}{

See the examples in \code{\link{mimic}}.
}

