% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimic.R
\name{mimic}
\alias{mimic}
\title{Modify a dataset to achieve certain sample summary statistics}
\usage{
mimic(x, x2)
}
\arguments{
\item{x}{A numeric matrix or data frame.  Each column contains observations
on a different variable.  Missing observations are not allowed.}

\item{which}{An integer in \{1, 2, 3, 4\}.  Which of Anscombe's dataset to
use.  Obviously, this makes very little difference.}
}
\value{
A dataset with the same format as \code{x}.  The returned dataset has the
same summary statistics as those in \code{stats}, except perhaps for the sample
size.
}
\description{
Modifies a dataset \code{x} so that it shares sample summary statistics with
another dataset \code{x2}.  Alternatively, the summary statistics themselves
}
\examples{
got_maps <- requireNamespace("maps", quietly = TRUE)
got_datasauRus <- requireNamespace("datasauRus", quietly = TRUE)
if (got_maps && got_datasauRus) {
  library(maps)
  library(datasauRus)
  dino <- datasaurus_dozen_wide[, c("dino_x", "dino_y")]
  UK <- mapdata("UK")
  new_UK <- mimic(UK, dino)
  plot(new_UK)
}
}
