% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.anscombe}
\alias{plot.anscombe}
\title{Plot diagnostics for an anscombe object}
\usage{
\method{plot}{anscombe}(x, old = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{'anscombe'}, a result of a call to \link{anscombise}
or \link{mimic}.}

\item{...}{Further arguments to be passed to \code{\link[graphics:plot.default]{plot}}}

\item{input}{A logical scalar.  Should the old, input data, that is, the
Anscombe's dataset chosen for \link{anscombise} or the argument \code{x2} to \link{mimic},
be plotted?  If \code{old = FALSE} then the new, output data are plotted.
If \code{old = TRUE} then the old data are plotted.}
}
\value{
Nothing is returned.
}
\description{
\code{plot} method for objects inheriting from class \code{"anscombe"}.
}
\details{
This function is only applicable in 2 or 3 dimensions, that is,
when \code{length(attr(x, "new_stats")$means)} is 2 or 3.
}
\section{Examples}{

See the examples in \link{anscombise} and \link{mimic}.
}

\seealso{
\link{anscombise} and \link{mimic}.
}
