% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anscombiser.R
\docType{data}
\name{trump}
\alias{trump}
\title{Donald Trump}
\format{
A matrix with 4885 rows and 2 columns: \code{x} and \code{y}.
}
\source{
This image was created by Accentaur from the Noun Project.
\url{https://thenounproject.com/term/donald-trump/727774/}
}
\usage{
trump
}
\description{
A dataset that provides an image of Donald Trump's face.
}
\keyword{datasets}
