\docType{data}
\name{anscombe}
\alias{anscombe}
\alias{anscombe1}
\alias{anscombe2}
\alias{anscombe3}
\alias{anscombe4}
\title{Anscombe's Quartet Separated}
\format{
All datasets are objects of class \code{data.frame} with 11 rows and 2
columns.
}
\source{
Anscombe's Quartet of 'Identical' Simple Linear Regressions:
\code{\link[datasets:anscombe]{datasets::anscombe}} in the \code{\link[datasets:datasets-package]{datasets}} package. The \code{i}th dataset is \code{datasets::anscombe[, c(i, i + 4)]}.
}
\usage{
anscombe1

anscombe2

anscombe3

anscombe4
}
\description{
Provides Anscombe's Quartet as separate data frames.
}
\references{
Anscombe, Francis J. (1973). Graphs in statistical analysis.
\emph{The American Statistician}, \strong{27}, 17-21. \doi{10.2307/2682899}
}
\keyword{datasets}
