% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anscombiser-internal.R
\name{anscombiser-internal}
\alias{anscombiser-internal}
\alias{is_wholenumber}
\title{Internal anscombiser functions}
\usage{
is_wholenumber(x, tol = .Machine$double.eps^0.5)
}
\description{
Internal anscombiser functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
