% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anscombiser-internal.R
\name{anscombiser-internal}
\alias{anscombiser-internal}
\alias{is_wholenumber}
\alias{is_pos_def}
\alias{minimal_rotation}
\alias{make_stats}
\title{Internal anscombiser functions}
\usage{
is_wholenumber(x, tol = .Machine$double.eps^0.5)

is_pos_def(x, tol = 1e-06)

minimal_rotation(x)

make_stats(x, stats, idempotent = FALSE)
}
\description{
Internal anscombiser functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
