% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.anscombe}
\alias{plot.anscombe}
\title{Plot diagnostics for an anscombe object}
\usage{
\method{plot}{anscombe}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"anscombe"}, a result of a call to
\code{\link{anscombise}} or \verb{\link{mimic}}.}

\item{...}{Further arguments to be passed to
\code{\link[graphics:plot.default]{plot}}.}
}
\value{
Nothing is returned.
}
\description{
\code{plot} method for objects inheriting from class \code{"anscombe"}.
}
\details{
This function is only applicable in 2 or 3 dimensions, that is,
when \code{length(attr(new_UK, "new_stats")$means)} is 2 or 3.
}
\section{Examples}{

See the examples in \code{\link{anscombise}} and \code{\link{mimic}}.
}

\seealso{
\code{\link{anscombise}} and \code{\link{mimic}}.
}
